requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
ccxt==1.84.15
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
pycares==4.1.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
scipy==1.8.1
six==1.16.0
urllib3==1.26.9
yarl==1.7.2
